+++ examples/basic_area.rb
@@ -8,7 +8,10 @@ window('Basic Area', 400, 400) {
   vertical_box {
     area {
       path { # a stable path is added declaratively
-        rectangle(0, 0, 400, 400)
+        figure(50, 50) {
+          rectangle(0, 0, 400, 400)
+          closed true
+        }
         
         fill r: 102, g: 102, b: 204, a: 1.0
       }